Get rid of the sleep dependency, not stable enough.
It gives so much trouble when installing.  Get rid of the sleep calll
is not too bad, as the loop does a sync HTTP request.  There is some
network I/O time to wait every time anyway (typically, with a very
good network, when ML is down, so on ECONNREFUSED, it takes a few ms
to do the loop).

Way better than the usability problem of using sleep.