Prepare command line parsing for the init command.
Support several named arguments for a command, at the end of the
command line (e.g. for `help`, the argument with name `cmd` is `xxx`
in `mlproj help xxx`).

Also support "rest" arguments.  Essentially the same, but all such
arguments are added to a list.  Needed for a command such as:

    mlproj init cluster host1 host2 host3